pyproject.toml
[project]
name = "olli"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bcrypt>=4.3.0",
    "fastapi>=0.115.12",
    "psycopg2-binary>=2.9.10",
    "python-decouple>=3.8",
    "sqlalchemy>=2.0.40",
]
